+++ Client-A/src/app/Components/Administrador/roles-list/roles-list.component.ts
@@ -1,6 +1,6 @@
 import { getTranslationDeclStmts } from '@angular/compiler/src/render3/view/template';
 import { Component, OnInit } from '@angular/core';
-import { RegistroRolService } from '../../Services/registro-rol.service';
+import { RegistroRolService } from '../../../Administrador/Services/registro-rol.service';
 
 import{AlertsService} from '../../../Services/alerts/alerts.service';
