+++ Client-A/src/app/Components/Editor/menu-principal-editor/menu-principal-editor.component.ts
