+++ Client-A/src/app/Models/archivo.ts
