+++ Client-A/src/app/Models/publicacion.ts
