+++ Client-A/src/app/Service/token-interceptor.service.ts
