+++ Client-A/src/app/app.module.ts
@@ -5,11 +5,11 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { RegistroUsuarioComponent } from './Administrador/Components/menu-principal/registro-usuario.component';
-import { UsuarioListComponent } from './Administrador/Components/usuario-list/usuario-list.component';
-import{RegistroUsuarioService} from './Administrador/Services/registro-usuario.service';
-import { RegistroArchivoService } from "./Editor/Services/registro-archivo.service";
-import { RolesListComponent } from './Administrador/Components/roles-list/roles-list.component';
+import { RegistroUsuarioComponent } from './Components/Administrador/menu-principal/registro-usuario.component';
+import { UsuarioListComponent } from './Components/Administrador/usuario-list/usuario-list.component';
+import{RegistroUsuarioService} from './Service/registro-usuario.service';
+import { RegistroArchivoService } from "./Service/registro-archivo.service";
+import { RolesListComponent } from './Components/Administrador/roles-list/roles-list.component';
 
 
 //imporar modulo para animacion de mensajes de alterta 
@@ -24,12 +24,12 @@ import { from } from 'rxjs';
 //importan el formulario modal
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 
-import { MenuPrincipalEditorComponent } from './Editor/Components/menu-principal-editor/menu-principal-editor.component';
+import { MenuPrincipalEditorComponent } from './Components/Editor/menu-principal-editor/menu-principal-editor.component';
  import { LoginComponent } from './Login/Components/login.component';
-import { PerfilesMujeresComponent } from './Editor/Components/perfiles-mujeres/perfiles-mujeres.component'
+import { PerfilesMujeresComponent } from './Components/Editor/perfiles-mujeres/perfiles-mujeres.component'
 import {LoginGuard} from './Login/Services/login.guard';
-import { NoticiasComponent } from './Editor/Components/noticias/noticias.component';
-import { MenuPublicacionComponent } from './Editor/Components/menu-publicacion/menu-publicacion.component'
+import { NoticiasComponent } from './Components/Editor/noticias/noticias.component';
+import { MenuPublicacionComponent } from './Components/Editor/menu-publicacion/menu-publicacion.component'
 // import {TokenInterceptorService} from './Login/Services/token-interceptor.service'
 
 @NgModule({
